tested album routes, fixed errors, done
